Remove extra arguments that aren't used.
According to the ExternalProject docs (1) `PREFIX` is used as the prefix
for the project directory generator expressions but is not otherwise
passed to the external project's build directly.

Similarly the docs for the `INSTALL_DIR` option state:

> Installation prefix to be placed in the <INSTALL_DIR> placeholder.
> This does not actually configure the external project to install to the given prefix.
> That must be done by passing appropriate arguments to the external project configuration step, e.g. using <INSTALL_DIR>.

So both of these arguments are unused and sources for potential
confusion.

(1) https://cmake.org/cmake/help/latest/module/ExternalProject.html

Signed-off-by: Steven! Ragnarök <steven@nuclearsandwich.com>