Remove unneeded BUILD_ALWAYS argument.
This argument does not do anything for this external project. Build farm
builds will always build this package when required since they run in a
clean workspace and since this external project is configured via git
CMake has the ability to detect changes and know that a rebuild is
required when working locally.

From the documentation for this option:

> Enabling this option forces the build step to always be run.
> This can be the easiest way to robustly ensure that the external project's own build dependencies are evaluated rather than relying on the default success timestamp-based method.
> This option is not normally needed unless developers are expected to modify something the external project's build depends on in a way that is not detectable via the step target dependencies (e.g. SOURCE_DIR is used without a download method and developers might modify the sources in SOURCE_DIR).

Signed-off-by: Steven! Ragnarök <steven@nuclearsandwich.com>