Do not export dependencies on the vendored project.
ament's dependency export system relies on adding corresponding exported
dependencies in the `package.xml` file.

It actually would be possible to use this export if an additional
`buildtool_export_depend` on `ament_cmake` was used so that all
downstream packages required `ament_cmake` to build using this package.

However, that's a bit heavy handed for a vendor package in particular
and it is a recommended practice that vendor packages be as transparent
as possible so that they can be removed once deprecated.

This change will require updates to packages consuming this vendor
package.

Signed-off-by: Steven! Ragnarök <steven@nuclearsandwich.com>